Use the generic event from netmsg to write text only
If we haven't registered a string resource library, the event will
show an error message on the front page and the text will only be
available in the details. This looks ugly. We use the same trick as
Apache and reuse the event ID 3299 from netmsg.dll, which consists
only of placeholders, to write the log text to the front page.